Move cached getter functions to model
